rename aggregatormanager
