+++ relay-metrics/src/lib.rs
@@ -68,15 +68,14 @@
     html_favicon_url = "https://raw.githubusercontent.com/getsentry/relay/master/artwork/relay-icon.png"
 )]
 
-mod aggregator;
+/// Only public for benchmarks.
+pub mod aggregator;
+
 mod aggregatorservice;
 mod bucket;
 mod protocol;
 mod router;
 mod statsd;
-
-// only public for benchmarks.
-pub use aggregator::*;
 
 pub use aggregatorservice::*;
 pub use bucket::*;
